MiniResearch/MiniResearchAgent/requirements.txt
numpy
pandas
faiss-cpu
sentence-transformers
PyMuPDF
streamlit
